make several components
